+++ assets/ts/models/PotentialGame.ts
@@ -1,0 +1,6 @@
+export class PotentialGame {
+	public commandLine: string[];
+	public uuid: string;
+
+	constructor(public name: string, public details?: any) {}
+}
